anki-ojad: rewrite to use static files
This plugin was terribly slow because it would query ojad for every
entry, freezing poor old python. Also html.parser is painful to work
with. I ended up:

  * replacing html.parser with elementtree
  * scraping all entries from ojad and storing into files
  * load dict from files when needed and use that instead of querying

There are still things left to do, like get the masu stem from verbs
(and probably lots of other verb conjugation things) but this is enough
for now.